requirements.txt
tensorflow>=2.13
keras>=2.13
streamlit
numpy
pillow
scikit-learn